wxWidgets: tweak last commit + squash a compiler warning
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@3377 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/corelib/PWSfileV3.cpp
@@ -820,7 +820,7 @@ int PWSfileV3::ReadHeader()
         for (int n = 0; n < num; n++) {
           uuid_array_t uuid;
           int nscanned = 0;
-          for (int i = 0; i < sizeof(uuid_array_t); i++) {
+          for (size_t i = 0; i < sizeof(uuid_array_t); i++) {
 #if (_MSC_VER >= 1400)
             nscanned += _stscanf_s(lpsz_string, _T("%02x"), &pfield[i]);
 #else
